requirements.txt
pandas
torch
scikit-learn
numpy
seaborn
google-genai
fastapi
uvicorn
pydantic
